gausian.ai.frontend/film-mvp/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=0.19.0
pydantic>=2.0.0

# LLM and AI
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
langchain>=0.0.300
langchain-openai>=0.0.1
langchain-google-genai>=0.0.1

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-multipart==0.0.6
opencv-python>=4.8.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# ComfyUI Integration
requests>=2.31.0
rich>=13.0.0
moviepy>=1.0.3
Pillow>=9.0.0
numpy>=1.21.0
